move searchable dropdown to molecules
